fix(token-generator): clean up oidc token files
